requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
pydantic-settings
python-dotenv
dnspython
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart>=0.0.9
